requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
pandas==2.1.4
python-multipart==0.0.6
jinja2==3.1.2
apscheduler==3.10.4
python-dotenv==1.0.0
alembic==1.12.1